Final Commit repost
(I moved the whole folder cuz it was in the wrong place)

Right, so I spoke to you today btw (Prof. Pippin) and you said I should write about some of my problems and creative solutions I found to them.

So one of my biggest problems was how I was going to check collision with the mouse to determine practically everything in the game.

What I ended up doing was creating if statements that checked whether the mouse was within a provided width and height. If that rectangular area of collision had to move, I would simply move it at the same rate as whatever image was moving with it.

Another, even BIGGER issue I had was with spawning enemies. This issue stemmed from a couple of things.

1. How would I tell my program when to create a specific enemy?
2. How would that enemy be continuously drawn while "alive" and continuously not drawn while "dead"?

I ended up solving both these issues through a system setInterval methods calling a "time" function every second. That would trigger a function with an if statement that was looking for time to go to or above 1. Inside this if statement was a boolean variable that would be set to true, which triggered a function. If I didn't have this variable, the function that I triggered would have triggered a very large amount of times due to the draw(); function behind everything.